feat: support decimal ratings (float64) in review API
- Changed Rating field from int to float64 in CreateReviewRequest and UpdateReviewRequest
- Updated CreateReview() and UpdateReview() usecase functions to accept float64
- Convert float64 to int when storing in database (e.g., 4.15 -> 4)
- Allows frontend to submit ratings like 4.15 via API

Fixes: POST /product/{id}/review/{reviewid} now accepts decimal ratings

+++ internal/interface/handler/productreview/handler.go
@@ -21,7 +21,7 @@ import (
 
 // Request/Response structures
 type CreateReviewRequest struct {
-	Rating            int             `json:"rating"`
+	Rating            float64         `json:"rating"`
 	Reviews           string          `json:"reviews"`
 	SourceURL         *string         `json:"source_url,omitempty"`
 	AdditionalDetails json.RawMessage `json:"additional_details,omitempty"`
@@ -29,7 +29,7 @@ type CreateReviewRequest struct {
 }
 
 type UpdateReviewRequest struct {
-	Rating            int             `json:"rating"`
+	Rating            float64         `json:"rating"`
 	Reviews           string          `json:"reviews"`
 	SourceURL         *string         `json:"source_url,omitempty"`
 	AdditionalDetails json.RawMessage `json:"additional_details,omitempty"`
